Basic Structure
This version is just the basic styling of the list. Its only functionality is when the checkbox is checked it changes the style of the checkbox. When there are multiple checkboxes with different states the style does not match the corresponding state of the checkbox. I also plan to add the ability to add and delete tasks.